Refactor. Move view files to new directory and Fix router import
